Ahora los usuarios mas nuevos tienen dos roles: ROLE_OWNER, ROLE_ADMIN Cuando se crea un user, este obtiene los privilegios de un admin, entonces todas aquellas paginas con ...:8080/admin/** son bloqueadas al menos que sean admin, Lo mismo para owner, toda cuenta que se cree desde el MENU: ...:8080/owner_signup tendra solo permisos de owner, es decir todo lo que comienze con ...:8080/owner/** solo esta disponible para ellos
Todo modulo que vayan a crear favor de poner el prefijo /owner/ o /admin/

+++ spring-petclinic-master/src/main/java/org/springframework/samples/petclinic/user/UserRepository.java
@@ -18,7 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
  *
  * @author AugustoRuCle
  */
-public interface UserRepository extends Repository<User, Integer>{
+public interface UserRepository extends Repository<User, Long>{
 
     @Query("SELECT user FROM User user WHERE user.id = :id")
     @Transactional(readOnly = true)
